Change "ReleaseNotes" to "ChangeLog"
Also add custom block to it directly instead of around it seperately in the Markdown file.
