+++ src/download/README.md
@@ -17,6 +17,4 @@ Latest release: **<VersionTag/>**
 
 <DownloadButtons downloadStableTag="Tachiyomi" downloadPreviewTag="Tachiyomi Preview"/>
 
-::: guide Release notes
-<ReleaseNotes/>
-:::
+<ChangeLog />
